Starting the command

+++ src/dotnet-toolbox/VersionFile/VersionFile.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using Newtonsoft.Json;
@@ -7,6 +8,10 @@ namespace DotNetToolbox.VersionMetadata
 {
     public class VersionFile
     {
+        public List<PackageMetadata> ReadFile(string versionFilePath)
+        {
+            throw new NotImplementedException();
+        }
 
         public static void WriteVersion(PackageMetadata pkg, string versionFilePath)
         {
@@ -21,6 +26,17 @@ namespace DotNetToolbox.VersionMetadata
                 _contents.Remove(existingVersion);
             }
             _contents.Add(pkg);
+            File.WriteAllText(versionFilePath, JsonConvert.SerializeObject(_contents));
+        }
+
+        public static void RemoveVersion(PackageMetadata pkg, string versionFilePath)
+        {
+            var _contents = new List<PackageMetadata>();
+            if (!File.Exists(versionFilePath))
+                throw new InvalidOperationException("WTF??");
+
+            var stuff = JsonConvert.DeserializeObject<List<PackageMetadata>>(File.ReadAllText(versionFilePath));
+            var newStuff = stuff.Where(p => p.PackageId == pkg.PackageId);
             File.WriteAllText(versionFilePath, JsonConvert.SerializeObject(_contents));
         }
